Add explicit ::numeric casts to fix PostgreSQL type inference error
PostgreSQL cannot deduce a consistent type when the same parameter ($2)
appears in assignment, arithmetic, and comparison contexts. Explicit
::numeric casts on all numeric parameters resolve the ambiguity.

+++ app/api/cuentas-pagar/[id]/route.ts
@@ -109,13 +109,13 @@ export async function PATCH(request: NextRequest, { params }: { params: Promise<
       try {
         await client.query(
           `UPDATE cuentas_pagar SET
-             monto_bs = $2,
-             monto_usd = CASE WHEN $5 > 0 THEN $2 / $5 ELSE monto_usd END,
-             monto_original_bs = COALESCE($3, monto_original_bs),
-             monto_pagado_bs = COALESCE(monto_pagado_bs, 0) + $4,
-             estado = CASE WHEN $2 <= 0 THEN 'PAGADO' ELSE 'PENDIENTE_PARCIAL' END,
+             monto_bs = $2::numeric,
+             monto_usd = CASE WHEN $5::numeric > 0 THEN $2::numeric / $5::numeric ELSE monto_usd END,
+             monto_original_bs = COALESCE($3::numeric, monto_original_bs),
+             monto_pagado_bs = COALESCE(monto_pagado_bs, 0) + $4::numeric,
+             estado = CASE WHEN $2::numeric <= 0 THEN 'PAGADO' ELSE 'PENDIENTE_PARCIAL' END,
              fecha_vencimiento = COALESCE($6::date, fecha_vencimiento),
-             pagado_at = CASE WHEN $2 <= 0 THEN NOW() ELSE NULL END
+             pagado_at = CASE WHEN $2::numeric <= 0 THEN NOW() ELSE NULL END
            WHERE id = $1`,
           [id, restanteBs, montoOriginalBs, montoPagadoBs, tasaDia, nuevaFechVenc]
         );
@@ -125,11 +125,11 @@ export async function PATCH(request: NextRequest, { params }: { params: Promise<
         await client.query("ROLLBACK TO SAVEPOINT sp_update");
         await client.query(
           `UPDATE cuentas_pagar SET
-             monto_bs = $2,
-             monto_usd = CASE WHEN $4 > 0 THEN $2 / $4 ELSE monto_usd END,
-             estado = CASE WHEN $2 <= 0 THEN 'PAGADO' ELSE 'PENDIENTE_PARCIAL' END,
+             monto_bs = $2::numeric,
+             monto_usd = CASE WHEN $4::numeric > 0 THEN $2::numeric / $4::numeric ELSE monto_usd END,
+             estado = CASE WHEN $2::numeric <= 0 THEN 'PAGADO' ELSE 'PENDIENTE_PARCIAL' END,
              fecha_vencimiento = COALESCE($5::date, fecha_vencimiento),
-             pagado_at = CASE WHEN $2 <= 0 THEN NOW() ELSE NULL END
+             pagado_at = CASE WHEN $2::numeric <= 0 THEN NOW() ELSE NULL END
            WHERE id = $1`,
           [id, restanteBs, montoPagadoBs, tasaDia, nuevaFechVenc]
         );
